+++ flake.nix
@@ -5,11 +5,11 @@
     nixpkgs.url = "github:nixos/nixpkgs/nixpkgs-22.11-darwin";
     cardano-wallet-unpatched.url = "github:cardano-foundation/cardano-wallet/v2025-03-31";
     cardano-wallet-unpatched.flake = false; # otherwise, +10k quadratic dependencies in flake.lock…
-    cardano-node-override.url = "github:IntersectMBO/cardano-node/10.2.1";
+    cardano-node-override.url = "github:IntersectMBO/cardano-node/10.5.1";
     cardano-node-override.flake = false;
-    cardano-playground.url = "github:input-output-hk/cardano-playground/d3322dce0ab1c00386adc93899aabe9252342b54";
+    cardano-playground.url = "github:input-output-hk/cardano-playground/56ebfef5595c43014029b039ade01b0ef06233e0";
     cardano-playground.flake = false; # otherwise, +9k dependencies in flake.lock…
-    cardano-shell.url = "github:input-output-hk/cardano-shell/0d1d5f036c73d18e641412d2c58d4acda592d493";
+    cardano-shell.url = "github:input-output-hk/cardano-shell/79f48aa3aa8007a1597cbedc22031eab1f05decd";
     cardano-shell.flake = false;
     tullia.url = "github:input-output-hk/tullia";
     tullia.flake = false; # otherwie, +1k dependencies in flake.lock…