+++ nix/internal/launcher-config.nix
@@ -43,7 +43,7 @@ let
   fromCardanoPlayground = envName: let
     originalFiles = builtins.path {
       name = "cardano-playground-config-${envName}";
-      path = cardano-playground + ("/static/book.play.dev.cardano.org/environments/" + envName);
+      path = cardano-playground + ("/docs/environments/" + envName);
     };
 
     originalNodeConfig = builtins.fromJSON (builtins.unsafeDiscardStringContext (
@@ -63,8 +63,11 @@ let
     cardanoEnv = {
       inherit nodeConfig;
       topologyFile = originalFiles + "/topology.json";
+      peerSnapshotFile = originalFiles + "/peer-snapshot.json";
       metadataUrl = tokenMetadataServers.${envName};
-    };
+    } // (let
+      checkpointsFile = originalFiles + "/checkpoints.json";
+    in if builtins.pathExists checkpointsFile then { inherit checkpointsFile; } else {});
   };
 
   dirSep = if os == "windows" then "\\" else "/";
@@ -267,6 +270,12 @@ let
       cp ${genesisFile} $out/genesis.json
       cp $nodeConfigPath $out/config.yaml
       cp $topologyFile $out/topology.yaml
+      ${lib.optionalString (envCfg ? peerSnapshotFile) ''
+        cp ${envCfg.peerSnapshotFile} $out/peer-snapshot.json
+      ''}
+      ${lib.optionalString (envCfg ? checkpointsFile) ''
+        cp ${envCfg.checkpointsFile} $out/checkpoints.json
+      ''}
       ${lib.optionalString (network == "selfnode") ''
         cp ${envCfg.delegationCertificate} $out/delegation.cert
         cp ${envCfg.signingKey} $out/signing.key